Propagate gevent poller errors
In case wait() raises an exception while waiting in the gevent poller,
cancel the gRPC call and propagate the error back to the application.

To achieve this, store the exception in Python's own exception
registers, return from the poller with GRPC_ERROR_CANCELLED to cancel
the gRPC call and check if an exception has occurred once control
returns from the gRPC core back to Python.

Signed-off-by: Kostis Lolos <klolos@arrikto.com>